+++ src/resources/views/element-types/create.blade.php
@@ -4,17 +4,17 @@
 <div class="container">
     <div class="row">
         <div class="col-sm-12">
-            <div class="panel panel-default">
+            <div class="card">
 
                 <form action="{{ route('pagebuilder::element-types.store') }}" role="form" method="POST"  enctype="multipart/form-data">
                     {{ csrf_field() }}
 
-                    <div class="panel-heading panel-default">
-                        <h3 class="panel-title">Elemente</h3>
+                    <div class="card-header">
+                        <h3>Elemente</h3>
                         @lang('pagebuilder::crud.create_headline')
                     </div>
 
-                    <div class="panel-body">
+                    <div class="card-body">
                         
                         @include('pagebuilder::notifications')
                         <div class="form-group">
@@ -36,7 +36,7 @@
                             
                     </div>
 
-                    <div class="panel-footer">
+                    <div class="card-footer">
 
                         <div class="row">