+++ src/resources/views/main.blade.php
@@ -4,11 +4,11 @@
 <div class="container">
     <div class="row">
         <div class="col-sm-12">
-            <div class="panel panel-default">
-                <div class="panel-heading">
-                    <h3 class="panel-title">Pagebuilder</h3>
+            <div class="card">
+                <div class="card-header">
+                    <h3>Pagebuilder</h3>
                 </div>
-                <div class="panel-body">
+                <div class="card-body">
                     <div class="row">
                         <div class="col-sm-6">
                             Elements: {{ $elements->count() }}