+++ changelog.md
@@ -1,5 +1,12 @@
 ## Version History
 
+### v. 1.0.3
+
+- removed all JS assets (different package now)
+- fixed lang prefixes
+- added install command
+- added seeders
+
 ### v. 0.1.1
 
 - updated structure